Перегруппирует редьюсеры, добавит reselector, обновит компоненты App, Main, GenresList, обновит тесты
