+++ src/components/genres-list/genres-list.test.js
@@ -6,43 +6,23 @@ import {GENRE_ALL} from "../../utils/const.js";
 
 describe(`Render GenresList`, () => {
   it(`Should GenresList render correctly`, () => {
-    const moviesList = [
-      {
-        genre: `Comedies`,
-      },
-      {
-        genre: `Crime`,
-      },
-      {
-        genre: `Documentary`,
-      },
-      {
-        genre: `Drama`,
-      },
-      {
-        genre: `Horror`,
-      },
-      {
-        genre: `Kids & Family`,
-      },
-      {
-        genre: `Romance`,
-      },
-      {
-        genre: `Sci-Fi`,
-      },
-      {
-        genre: `Thrillers`,
-      },
-      {
-        genre: `Drama`,
-      }
+    const genresList = [
+      `Comedies`,
+      `Crime`,
+      `Documentary`,
+      `Drama`,
+      `Horror`,
+      `Kids & Family`,
+      `Romance`,
+      `Sci-Fi`,
+      `Thrillers`,
+      `Drama`,
     ];
 
     const tree = renderer
       .create(
           <GenresList
-            moviesList = {moviesList}
+            genresList = {genresList}
             currentGenre = {GENRE_ALL}
             onGenreItemClick = {() => {}}
           />)
