+++ src/reducer/data/data.test.js
@@ -5,8 +5,13 @@ import {createAPI} from "../../api.js";
 const api = createAPI(() => {});
 import {
   normalizeMovieData,
-  normalizeMoviesData
+  normalizeMoviesData,
+  getGenresList
 } from "../../utils/utils.js";
+import {
+  GENRE_ALL,
+  SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT
+} from "../../utils/const.js";
 
 const film = {
   name: `The Grand Budapest Hotel`,
@@ -23,6 +28,48 @@ const film = {
   genre: `Drama`,
   released: 2014,
 };
+const comments = [
+  {
+    id: 1,
+    userId: 4,
+    userName: `Kate Muir`,
+    rating: 8.9,
+    comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+    date: `2016-12-24`,
+  },
+  {
+    id: 2,
+    userId: 4,
+    userName: `Kate Muir`,
+    rating: 8.9,
+    comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+    date: `2016-12-24`,
+  },
+  {
+    id: 3,
+    userId: 4,
+    userName: `Kate Muir`,
+    rating: 8.9,
+    comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+    date: `2016-12-24`,
+  },
+  {
+    id: 4,
+    userId: 4,
+    userName: `Kate Muir`,
+    rating: 8.9,
+    comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+    date: `2016-12-24`,
+  },
+  {
+    id: 5,
+    userId: 4,
+    userName: `Kate Muir`,
+    rating: 8.9,
+    comment: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
+    date: `2016-12-24`,
+  },
+];
 const films = [
   {
     name: `Fantastic Beasts: The Crimes of Grindelwald`,
@@ -199,6 +246,12 @@ describe(`Data Reducer tests`, () => {
     expect(reducer(void 0, {})).toEqual({
       promoMovieCard: initialMovie,
       moviesList: [initialMovie],
+
+      genresList: [GENRE_ALL],
+      currentGenre: GENRE_ALL,
+      showedItemsInMoviesList: SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT,
+      currentMovie: null,
+      currentMovieComments: null,
     });
   });
 
@@ -221,6 +274,51 @@ describe(`Data Reducer tests`, () => {
       payload: films,
     })).toEqual({
       moviesList: films,
+      genresList: getGenresList(films),
+    });
+  });
+
+  it(`Reducer should change currentGenre to Drama`, () => {
+    expect(reducer({
+      currentGenre: GENRE_ALL,
+    }, {
+      type: ActionType.CHANGE_GENRE,
+      payload: `Drama`,
+    })).toEqual({
+      currentGenre: `Drama`,
+    });
+  });
+
+  it(`Reducer should change currentMovie`, () => {
+    expect(reducer({
+      currentMovie: null,
+      currentMovieComments: null,
+    }, {
+      type: ActionType.CHANGE_CURRENT_MOVIE,
+      payload: film,
+    })).toEqual({
+      currentMovie: film,
+      currentMovieComments: comments,
+    });
+  });
+
+  it(`Reducer should increase showedItemsInMoviesList`, () => {
+    expect(reducer({
+      showedItemsInMoviesList: SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT,
+    }, {
+      type: ActionType.SHOW_MORE_ITEMS_IN_MOVIES_LIST,
+    })).toEqual({
+      showedItemsInMoviesList: SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT + SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT,
+    });
+  });
+
+  it(`Reducer should reset showedItemInMoviesList to default`, () => {
+    expect(reducer({
+      showedItemsInMoviesList: SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT + SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT,
+    }, {
+      type: ActionType.RESET_SHOWED_ITEMS_IN_MOVIES_LIST,
+    })).toEqual({
+      showedItemsInMoviesList: SHOWED_ITEMS_IN_MOVIES_LIST_DEFAULT,
     });
   });
 });
@@ -237,6 +335,32 @@ describe(`Data ActionCreator tests`, () => {
     expect(ActionCreator.loadMoviesList(films)).toEqual({
       type: ActionType.LOAD_MOVIES_LIST,
       payload: films,
+    });
+  });
+
+  it(`ActionCreator for changeGenre return correct action`, () => {
+    expect(ActionCreator.changeGenre(`Drama`)).toEqual({
+      type: ActionType.CHANGE_GENRE,
+      payload: `Drama`,
+    });
+  });
+
+  it(`ActionCreator for changeCurrentMovie return correct action`, () => {
+    expect(ActionCreator.changeCurrentMovie(film)).toEqual({
+      type: ActionType.CHANGE_CURRENT_MOVIE,
+      payload: film,
+    });
+  });
+
+  it(`ActionCreator for increase showedItemInMoviesList return correct action`, () => {
+    expect(ActionCreator.showMoreItemsInMoviesList()).toEqual({
+      type: ActionType.SHOW_MORE_ITEMS_IN_MOVIES_LIST,
+    });
+  });
+
+  it(`ActionCreator for return defautl value showedItemInMoviesList return correct action`, () => {
+    expect(ActionCreator.resetShowedItemsInMoviesList()).toEqual({
+      type: ActionType.RESET_SHOWED_ITEMS_IN_MOVIES_LIST,
     });
   });
 });
